refactor(currencies): Transaction型を共有types.tsに抽出し重複定義を解消
- currencies/utils/types.tsにTransactionインターフェースを新設
- use-currencies.tsとtransaction-row.tsxの二重定義を統合
- isNavigableの冗長な!!tx.sessionIdチェックを除去
- handleNavigateToSessionの到達不能なデッドガードを削除
