+++ apps/web/src/features/currencies/pages/currency-detail-page/transaction-list/transaction-row/index.ts
@@ -1,1 +1,2 @@
-export { type Transaction, TransactionRow } from "./transaction-row";
+export { type Transaction } from "@/features/currencies/utils/types";
+export { TransactionRow } from "./transaction-row";
